[#551] Small fixes and formatting
