+++ app/build.gradle
@@ -95,7 +95,7 @@ android.applicationVariants.all { variant ->
 }
 
 dependencies {
-    compile fileTree(include: '*.jar', dir: 'libs')
+    compile fileTree(dir: 'libs', include: '*.jar')
     compile 'com.android.support:appcompat-v7:21.0.0'
     compile 'com.android.support:support-annotations:21.0.0'
     compile 'com.google.android.gms:play-services-base:7.5.0'
@@ -114,5 +114,4 @@ dependencies {
     testCompile 'org.powermock:powermock-classloading-xstream:1.6.2'
     androidTestCompile 'com.google.dexmaker:dexmaker:1.2'
     androidTestCompile 'com.google.dexmaker:dexmaker-mockito:1.2'
-    compile files('libs/sentry-1.1.4.jar')
 }
