+++ app/src/main/java/org/akvo/flow/activity/SurveyActivity.java
@@ -59,8 +59,6 @@ import org.akvo.flow.util.ViewUtil;
 
 import java.lang.ref.WeakReference;
 
-import timber.log.Timber;
-
 public class SurveyActivity extends ActionBarActivity implements RecordListListener,
         DrawerFragment.DrawerListener, DatapointsFragment.DatapointFragmentListener {
     private static final String TAG = SurveyActivity.class.getSimpleName();
@@ -154,7 +152,6 @@ public class SurveyActivity extends ActionBarActivity implements RecordListListe
         if (savedInstanceState == null) {
             displaySelectedUser();
         }
-        Timber.e(new Exception("SurveyActivity"), "Testing raven ");
     }
 
     private void initializeToolBar() {
