+++ app/src/main/java/org/akvo/flow/util/logging/RavenTree.java
@@ -24,10 +24,11 @@ import com.getsentry.raven.android.Raven;
 
 import timber.log.Timber;
 
-public class RavenTree extends Timber.Tree {
+class RavenTree extends Timber.Tree {
 
     @Override
-    protected void log(int priority, @Nullable String tag, @Nullable String message, @Nullable Throwable t) {
+    protected void log(int priority, @Nullable String tag, @Nullable String message,
+            @Nullable Throwable t) {
         //Do not send reports for verbose, debug or info
         if (priority == Log.VERBOSE || priority == Log.DEBUG || priority == Log.INFO) {
             return;
